refactor(mysql): move metadata SQL into sql module
