+++ src/infra/adapters/mysql/metadata/mod.rs
@@ -10,9 +10,7 @@ mod preview;
 mod signature;
 mod table_detail;
 
-pub(super) use preview::{
-    build_preview_query, convert_preview_values, fetch_preview_metadata, preview_result_columns,
-};
+pub(super) use preview::{convert_preview_values, fetch_preview_metadata, preview_result_columns};
 
 #[async_trait]
 impl MetadataProvider for MySqlAdapter {
