+++ src/infra/adapters/mysql/sql/mod.rs
@@ -5,5 +5,12 @@ mod grid_write;
 mod literal;
 mod metadata;
 
-pub(super) use literal::{quote_identifier, quote_string};
-pub(super) use metadata::build_metadata_select_query;
+pub(super) use metadata::{
+    COLUMN_METADATA_RESULT_COLUMNS, FOREIGN_KEY_RESULT_COLUMNS, INDEX_RESULT_COLUMNS,
+    SIGNATURE_COLUMNS_QUERY, SIGNATURE_COLUMNS_RESULT_COLUMNS, SIGNATURE_FOREIGN_KEYS_QUERY,
+    SIGNATURE_UNIQUE_COLUMNS_QUERY, SIGNATURE_UNIQUE_COLUMNS_RESULT_COLUMNS, TABLES_QUERY,
+    TABLES_RESULT_COLUMNS, TRIGGER_RESULT_COLUMNS, UNIQUE_COLUMN_RESULT_COLUMNS,
+    build_metadata_select_query, build_preview_query, columns_query, foreign_keys_query,
+    indexes_query, preview_identity_alias, show_create_query, show_create_result_columns,
+    table_query, triggers_query, unique_columns_query,
+};
